compute: rearrange utils
Shared module grew over time and many of the classes and constants
should not be there.
